Fixed the order of the rules to 1- install/update all packages 2- generate all typings 3- compile all Jakefile.ts files
Make the path of the commands absolute, so that they can be used from anywhere
Add a flag in makefile and jakefile to turn logging on/off. Other makefiles can overwrite this variable.
Make the nodejs version so that other makefiles can overwrite it (should be use with caution when jakets is shared among multiple projects)

with these changes, we can handle jakets included in the package.json of other projects.